third_party/blink/web_tests/external/wpt/encoding/legacy-mb-schinese/gb18030/gb18030-decoder.any-expected.txt

This is a testharness.js-based test.
Found 18 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] gb18030 decoder: GB18030-2022 19
  assert_equals: expected "йҫҙ" but got "о һ"
[FAIL] gb18030 decoder: GB18030-2022 20
  assert_equals: expected "йҫө" but got "о Ұ"
[FAIL] gb18030 decoder: GB18030-2022 21
  assert_equals: expected "йҫ¶" but got "о «"
[FAIL] gb18030 decoder: GB18030-2022 22
  assert_equals: expected "йҫ·" but got "о ¬"
[FAIL] gb18030 decoder: GB18030-2022 23
  assert_equals: expected "йҫё" but got "о І"
[FAIL] gb18030 decoder: GB18030-2022 24
  assert_equals: expected "йҫ№" but got "оЎғ"
[FAIL] gb18030 decoder: GB18030-2022 25
  assert_equals: expected "йҫә" but got "оЎ”"
[FAIL] gb18030 decoder: GB18030-2022 26
  assert_equals: expected "йҫ»" but got "оЎӨ"
[FAIL] gb18030 decoder: GB18030-2022 27
  assert_equals: expected "пёҗ" but got "оһҚ"
[FAIL] gb18030 decoder: GB18030-2022 28
  assert_equals: expected "пё‘" but got "оһҸ"
[FAIL] gb18030 decoder: GB18030-2022 29
  assert_equals: expected "пё’" but got "оһҺ"
[FAIL] gb18030 decoder: GB18030-2022 30
  assert_equals: expected "пё“" but got "оһҗ"
[FAIL] gb18030 decoder: GB18030-2022 31
  assert_equals: expected "пё”" but got "оһ‘"
[FAIL] gb18030 decoder: GB18030-2022 32
  assert_equals: expected "пё•" but got "оһ’"
[FAIL] gb18030 decoder: GB18030-2022 33
  assert_equals: expected "пё–" but got "оһ“"
[FAIL] gb18030 decoder: GB18030-2022 34
  assert_equals: expected "пё—" but got "оһ”"
[FAIL] gb18030 decoder: GB18030-2022 35
  assert_equals: expected "пёҳ" but got "оһ•"
[FAIL] gb18030 decoder: GB18030-2022 36
  assert_equals: expected "пёҷ" but got "оһ–"
Harness: the test ran to completion.
